version 1.0/backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.0.5
SQLAlchemy==2.0.23
jieba==0.42.1
snownlp==0.12.3
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.2
matplotlib==3.7.2
wordcloud==1.9.2
transformers==4.35.0
torch==2.1.0
Pillow==10.1.0
python-dotenv==1.0.0
Werkzeug==2.3.7
pypinyin==0.49.0
requests==2.31.0
openai==1.3.0
anthropic==0.7.0

